Refs #188 - created SecurityContextMockInjector
